moved wait_for inside remote actor, tests pass now
